Move Context to context.go.
